+++ src/example/java/io/github/computerdaddyguy/jfiletreeprettyprinter/example/Filtering.java
@@ -6,8 +6,9 @@ import io.github.computerdaddyguy.jfiletreeprettyprinter.PathPredicates;
 public class Filtering {
 
 	public static void main(String[] args) {
+		var hasJavaExtensionPredicate = PathPredicates.builder().hasExtension("java").build();
 		var prettyPrinter = FileTreePrettyPrinter.builder()
-			.customizeOptions(options -> options.filter(PathPredicates.hasExtension("java")))
+			.customizeOptions(options -> options.filter(hasJavaExtensionPredicate))
 			.build();
 
 		var tree = prettyPrinter.prettyPrint("src/example/resources/filtering");