+++ src/test/java/io/github/computerdaddyguy/jfiletreeprettyprinter/LineExtensionTest.java
@@ -30,16 +30,16 @@ class LineExtensionTest {
 	void example_dir_match() {
 
 		Function<Path, String> lineExtension = path -> {
-			if (PathUtils.isDirectory(path) && PathUtils.hasName(path, "api")) {
+			if (PathPredicates.hasFullPathMatchingGlob(path, "**/src/main/java/api")) {
 				return "\t\t\t// All API code: controllers, etc.";
 			}
-			if (PathUtils.isDirectory(path) && PathUtils.hasName(path, "domain")) {
+			if (PathPredicates.hasFullPathMatchingGlob(path, "**/src/main/java/domain")) {
 				return "\t\t\t// All domain code: value objects, etc.";
 			}
-			if (PathUtils.isDirectory(path) && PathUtils.hasName(path, "infra")) {
+			if (PathPredicates.hasFullPathMatchingGlob(path, "**/src/main/java/infra")) {
 				return "\t\t\t// All infra code: database, email service, etc.";
 			}
-			if (PathUtils.isFile(path) && PathUtils.hasName(path, "application.properties")) {
+			if (PathPredicates.hasNameMatchingGlob(path, "*.properties")) {
 				return "\t// Config file";
 			}
 			return null;
@@ -70,7 +70,7 @@ class LineExtensionTest {
 	void compact_dir_first_dir() {
 
 		Function<Path, String> lineExtension = p -> {
-			if (PathUtils.hasName(p, "dirA")) {
+			if (PathPredicates.hasName(p, "dirA")) {
 				return " // 1";
 			}
 			return null;
@@ -92,7 +92,7 @@ class LineExtensionTest {
 	void compact_dir_middle_dir() {
 
 		Function<Path, String> lineExtension = p -> {
-			if (PathUtils.hasName(p, "dirB")) {
+			if (PathPredicates.hasName(p, "dirB")) {
 				return " // 2";
 			}
 			return null;
@@ -114,7 +114,7 @@ class LineExtensionTest {
 	void compact_dir_last_dir() {
 
 		Function<Path, String> lineExtension = p -> {
-			if (PathUtils.hasName(p, "dirC")) {
+			if (PathPredicates.hasName(p, "dirC")) {
 				return " // 3";
 			}
 			return null;